refactor: move group management and components into separate directory
