+++ src/components/messenger/group-management/index.test.tsx
@@ -1,11 +1,11 @@
 import { shallow } from 'enzyme';
 
 import { GroupManagement, Properties } from '.';
-import { Stage } from '../../../../store/group-management';
-import { AddMembersPanel } from '../add-members-panel';
-import { EditConversationPanel } from '../edit-conversation-panel';
-import { EditConversationState } from '../../../../store/group-management/types';
-import { ViewGroupInformationPanel } from '../view-group-information-panel';
+import { Stage } from '../../../store/group-management';
+import { AddMembersPanel } from './add-members-panel';
+import { EditConversationPanel } from './edit-conversation-panel';
+import { EditConversationState } from '../../../store/group-management/types';
+import { ViewGroupInformationPanel } from './view-group-information-panel';
 
 describe(GroupManagement, () => {
   const subject = (props: Partial<Properties>) => {
