+++ src/components/messenger/group-management/view-group-information-panel/index.test.tsx
